requirements.txt
arrow==1.3.0
asgiref==3.8.1
autopep8==2.3.1
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
distlib==0.3.8
Django==4.2.13
django-cleanup==9.0.0
django-filter==24.3
django-multiselectfield==0.1.13
django-simpleui==2024.4.1
djangorestframework==3.15.2
et-xmlfile==1.1.0
filelock==3.15.4
idna==3.8
importlib_metadata==8.5.0
Jinja2==3.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
pwlf==2.2.1
pycodestyle==2.12.1
pyDOE==0.3.8
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scipy==1.10.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tomli==2.0.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uWSGI==2.0.26
virtualenv==20.26.3
zipp==3.20.2
